adds refractored code

+++ crud/mysql/function.php
@@ -3,6 +3,7 @@
 //create user
 
 function createUser(){
+if(isset($_POST['submit'])){
   global $connection;
     if (isset($_POST['submit'])){
         $username = $_POST['username'];
@@ -18,6 +19,7 @@ function createUser(){
         }
 
     }
+  }
 }
 
 
@@ -25,62 +27,64 @@ function createUser(){
 
 // show data
 function showAllData(){
-  global $connection;
+    global $connection;
 
-  $query = "SELECT * FROM users";
+    $query = "SELECT * FROM users";
 
-  $result = mysqli_query($connection, $query);
-  if ($result){
-      
-  } else {
-      die("Query failed" . mysqli_error($connection));
-  }
+    $result = mysqli_query($connection, $query);
+    if ($result){
+        
+    } else {
+        die("Query failed" . mysqli_error($connection));
+    }
 
 
-  while($row = mysqli_fetch_assoc($result)){
-      $id = $row['id'];
-      echo "<option value='".$id."'>".$id."</option>";
-  }
+    while($row = mysqli_fetch_assoc($result)){
+        $id = $row['id'];
+        echo "<option value='".$id."'>".$id."</option>";
+    }
 }
-
 
 //update
 
 function updateTable(){
-  global $connection;
-  $username = $_POST['username'];
-  $password = $_POST['password'];
-  $id = $_POST['id'];
+  if(isset($_POST['submit'])){
+    global $connection;
+    $username = $_POST['username'];
+    $password = $_POST['password'];
+    $id = $_POST['id'];
 
-  $query = "UPDATE users SET ";
-  $query .= "username = '$username', ";
-  $query .= "password = '$password' ";
-  $query .= "WHERE id = ".$id." ";
+    $query = "UPDATE users SET ";
+    $query .= "username = '$username', ";
+    $query .= "password = '$password' ";
+    $query .= "WHERE id = ".$id." ";
 
-  $result = mysqli_query($connection, $query);
-  if(!$result){
-      die("Query failed" . mysqli_error($connection));
-  } else {
-      echo 'happpy ahyaplsdkjf';
+    $result = mysqli_query($connection, $query);
+    if(!$result){
+        die("Query failed" . mysqli_error($connection));
+    } else {
+        echo 'record updated';
+    }
   }
 }
 
 function deleteUser(){
-  global $connection;
-  $username = $_POST['username'];
-  $password = $_POST['password'];
-  $id = $_POST['id'];
+  if(isset($_POST['submit'])){  
+    global $connection;
+    $username = $_POST['username'];
+    $password = $_POST['password'];
+    $id = $_POST['id'];
 
-  $query = "DELETE FROM users ";
-  $query .= "WHERE id = ".$id." ";
+    $query = "DELETE FROM users ";
+    $query .= "WHERE id = ".$id." ";
 
-  $result = mysqli_query($connection, $query);
-  if(!$result){
-      die("Query failed" . mysqli_error($connection));
-  } else {
-      echo 'happpy ahyaplsdkjf';
+    $result = mysqli_query($connection, $query);
+    if(!$result){
+        die("Query failed" . mysqli_error($connection));
+    } else {
+        echo 'Record Deleted';
+    }
   }
 }
-
 
 ?>
